+++ flight/PiOS/STM32F30x/pios_internal_adc.c
@@ -30,14 +30,14 @@
 
 #if defined(PIOS_INCLUDE_ADC)
 
-static void PIOS_INTERNAL_ADC_PinConfig(uint32_t internal_adc_id);
-static void PIOS_INTERNAL_DMAConfig(uint32_t internal_adc_id);
-int32_t PIOS_INTERNAL_ADC_Init(uint32_t *internal_adc_id, const struct pios_internal_adc_cfg *cfg);
-static void PIOS_INTERNAL_ADC_Converter_Config(uint32_t internal_adc_id);
-static bool PIOS_INTERNAL_ADC_Available(uint32_t internal_adc_id, uint32_t pin);
-static int32_t PIOS_INTERNAL_ADC_PinGet(uint32_t internal_adc_id, uint32_t pin);
-static uint8_t PIOS_INTERNAL_ADC_NumberOfChannels(uint32_t internal_adc_id);
-static float PIOS_INTERNAL_ADC_LSB_Voltage(uint32_t internal_adc_id);
+static void PIOS_INTERNAL_ADC_PinConfig(uintptr_t internal_adc_id);
+static void PIOS_INTERNAL_DMAConfig(uintptr_t internal_adc_id);
+int32_t PIOS_INTERNAL_ADC_Init(uintptr_t *internal_adc_id, const struct pios_internal_adc_cfg *cfg);
+static void PIOS_INTERNAL_ADC_Converter_Config(uintptr_t internal_adc_id);
+static bool PIOS_INTERNAL_ADC_Available(uintptr_t internal_adc_id, uint32_t pin);
+static int32_t PIOS_INTERNAL_ADC_PinGet(uintptr_t internal_adc_id, uint32_t pin);
+static uint8_t PIOS_INTERNAL_ADC_NumberOfChannels(uintptr_t internal_adc_id);
+static float PIOS_INTERNAL_ADC_LSB_Voltage(uintptr_t internal_adc_id);
 
 const struct pios_adc_driver pios_internal_adc_driver = {
 		.available = PIOS_INTERNAL_ADC_Available,
@@ -108,7 +108,7 @@ static struct pios_internal_adc_dev * PIOS_INTERNAL_ADC_Allocate()
  * @brief Configures the pins used on the ADC device
  * \param[in] handle to the ADC device
  */
-static void PIOS_INTERNAL_ADC_PinConfig(uint32_t internal_adc_id)
+static void PIOS_INTERNAL_ADC_PinConfig(uintptr_t internal_adc_id)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 	if (!PIOS_INTERNAL_ADC_validate(adc_dev)) {
@@ -131,7 +131,7 @@ static void PIOS_INTERNAL_ADC_PinConfig(uint32_t internal_adc_id)
  * @brief Configures the DMA used on the ADC device
  * \param[in] handle to the ADC device
  */
-static void PIOS_INTERNAL_DMAConfig(uint32_t internal_adc_id)
+static void PIOS_INTERNAL_DMAConfig(uintptr_t internal_adc_id)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 	if (!PIOS_INTERNAL_ADC_validate(adc_dev)) {
@@ -201,7 +201,7 @@ static void PIOS_INTERNAL_DMAConfig(uint32_t internal_adc_id)
  * @brief Configures the ADC device
  * \param[in] handle to the ADC device
  */
-static void PIOS_INTERNAL_ADC_Converter_Config(uint32_t internal_adc_id)
+static void PIOS_INTERNAL_ADC_Converter_Config(uintptr_t internal_adc_id)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 
@@ -342,7 +342,7 @@ static void PIOS_INTERNAL_ADC_Converter_Config(uint32_t internal_adc_id)
 /**
  * @brief Init the ADC.
  */
-int32_t PIOS_INTERNAL_ADC_Init(uint32_t * internal_adc_id, const struct pios_internal_adc_cfg * cfg)
+int32_t PIOS_INTERNAL_ADC_Init(uintptr_t * internal_adc_id, const struct pios_internal_adc_cfg * cfg)
 {
 	PIOS_DEBUG_Assert(internal_adc_id); PIOS_DEBUG_Assert(cfg);
 
@@ -352,7 +352,7 @@ int32_t PIOS_INTERNAL_ADC_Init(uint32_t * internal_adc_id, const struct pios_int
 		return -1;
 	adc_dev->cfg = cfg;
 
-	*internal_adc_id = (uint32_t) adc_dev;
+	*internal_adc_id = (uintptr_t) adc_dev;
 	adc_dev->number_used_master_channels = 0;
 	adc_dev->number_used_slave_channels = 0;
 
@@ -399,11 +399,11 @@ int32_t PIOS_INTERNAL_ADC_Init(uint32_t * internal_adc_id, const struct pios_int
 	driver_instances[current_instances] = adc_dev;
 	++current_instances;
 
-	PIOS_INTERNAL_ADC_PinConfig((uint32_t) adc_dev);
+	PIOS_INTERNAL_ADC_PinConfig((uintptr_t) adc_dev);
 
-	PIOS_INTERNAL_DMAConfig((uint32_t) adc_dev);
+	PIOS_INTERNAL_DMAConfig((uintptr_t) adc_dev);
 
-	PIOS_INTERNAL_ADC_Converter_Config((uint32_t) adc_dev);
+	PIOS_INTERNAL_ADC_Converter_Config((uintptr_t) adc_dev);
 	return 0;
 }
 
@@ -413,7 +413,7 @@ int32_t PIOS_INTERNAL_ADC_Init(uint32_t * internal_adc_id, const struct pios_int
  * @param[in] internal_adc_id handler to the device to check
  * @return True if the pin is available.
  */
-static bool PIOS_INTERNAL_ADC_Available(uint32_t internal_adc_id, uint32_t pin)
+static bool PIOS_INTERNAL_ADC_Available(uintptr_t internal_adc_id, uint32_t pin)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 	/* Check if pin exists */
@@ -428,7 +428,7 @@ static bool PIOS_INTERNAL_ADC_Available(uint32_t internal_adc_id, uint32_t pin)
  * @param[in] internal_adc_id handler to the device to check
  * @return number of channels on the device.
  */
-static uint8_t PIOS_INTERNAL_ADC_NumberOfChannels(uint32_t internal_adc_id)
+static uint8_t PIOS_INTERNAL_ADC_NumberOfChannels(uintptr_t internal_adc_id)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 	if (!PIOS_INTERNAL_ADC_validate(adc_dev))
@@ -443,7 +443,7 @@ static uint8_t PIOS_INTERNAL_ADC_NumberOfChannels(uint32_t internal_adc_id)
  * @return ADC pin value averaged over the set of samples since the last reading.
  * @return -1 if pin doesn't exist
  */
-static int32_t PIOS_INTERNAL_ADC_PinGet(uint32_t internal_adc_id, uint32_t pin)
+static int32_t PIOS_INTERNAL_ADC_PinGet(uintptr_t internal_adc_id, uint32_t pin)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 	int32_t result;
@@ -533,7 +533,7 @@ static void PIOS_ADC_DMA_Handler(struct pios_internal_adc_dev *adc_dev)
 /**
  * @brief Gets the least significant bit voltage of the ADC
  */
-static float PIOS_INTERNAL_ADC_LSB_Voltage(uint32_t internal_adc_id)
+static float PIOS_INTERNAL_ADC_LSB_Voltage(uintptr_t internal_adc_id)
 {
 	struct pios_internal_adc_dev * adc_dev = (struct pios_internal_adc_dev *) internal_adc_id;
 	if (!PIOS_INTERNAL_ADC_validate(adc_dev)) {
